chore: move folders
To make way for new implementation
