+++ src/create-plugin-cli/createPluginCommand.spec.ts
