+++ src/create-plugin-cli/listPluginsCommand.ts
