+++ src/create-plugin-cli/loadPlugins.spec.ts
