+++ src/create-plugin-cli/pluginsCommand.spec.ts
