+++ src/create-plugin-cli/pluginsCommand.ts
