+++ src/create-plugin-cli/searchPackageCommand.spec.ts
