+++ src/create-plugin-cli/searchPackageCommand.ts
