+++ src/create-plugin-cli/searchPluginsCommand.spec.ts
