+++ src/create-plugin-cli/searchPluginsCommand.ts
