+++ src/index.ts
@@ -1,9 +1,9 @@
 import './enquirer-types'
 
 export { CliArgs, CliArgsWithoutDefaults, Parsable } from './argv-parser/types'
-export * from './cli'
+export * from './create-cli'
 export * from './errors'
-export * from './plugin-cli'
+export * from './create-plugin-cli'
 export * from './presenter'
 export * from './test-util'
