+++ src/test-util/argCommand.ts
@@ -1,4 +1,4 @@
-import { createCommand } from '../cli'
+import { createCommand } from '../create-cli'
 
 export const argCommand = createCommand({
   name: 'arg',
