+++ src/test-util/booleanOptionsCommand.ts
@@ -1,4 +1,4 @@
-import { createCommand } from '../cli'
+import { createCommand } from '../create-cli'
 
 export const booleanOptionsCommand = createCommand({
   name: 'opt',
