+++ src/test-util/contextCommand.ts
@@ -1,4 +1,4 @@
-import { createCommand } from '../cli'
+import { createCommand } from '../create-cli'
 
 export const contextCommand = createCommand({
   name: 'context',
