+++ src/test-util/echoCommand.ts
@@ -1,4 +1,4 @@
-import { createCommand } from '../cli'
+import { createCommand } from '../create-cli'
 
 export const echoCommand = createCommand({
   name: 'echo',
