+++ src/test-util/echoNameOptionCommand.ts
@@ -1,4 +1,4 @@
-import { createCommand } from '../cli'
+import { createCommand } from '../create-cli'
 
 export const echoNameOptionCommand = createCommand({
   name: 'echoNameOption',
