+++ src/test-util/helloCommand.ts
@@ -1,4 +1,4 @@
-import { createCommand } from '../cli'
+import { createCommand } from '../create-cli'
 
 export const helloCommand = createCommand({
   name: 'hello',
