+++ src/test-util/noopCommand.ts
@@ -1,4 +1,4 @@
-import { createCommand } from '../cli'
+import { createCommand } from '../create-cli'
 
 // istanbul ignore next
 export const noopCommand = createCommand({
