+++ src/test-util/numberOptionCommand.ts
@@ -1,4 +1,4 @@
-import { createCommand } from '../cli'
+import { createCommand } from '../create-cli'
 
 export const numberOptionCommand = createCommand({
   name: 'number-option',
