+++ src/test-util/stringOptionCommand.ts
@@ -1,4 +1,4 @@
-import { createCommand } from '../cli'
+import { createCommand } from '../create-cli'
 
 export const stringOptionCommand = createCommand({
   name: 'opt',
