+++ src/test-util/verboseCommand.ts
@@ -1,4 +1,4 @@
-import { createCommand } from '../cli'
+import { createCommand } from '../create-cli'
 
 export const verboseCommand = createCommand({
   name: 'verbose',
